move task definitions to a separate file to be reused by other DAGs #734
